requirements.txt
aiohttp>=3.9.0
aiohttp_jinja2>=1.6
pyrogram>=2.0.106
tgcrypto>=1.2.5
cryptography>=42.0.0
uvloop>=0.19.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
pytz
